renaming
* rename `Norton::Hash` to `Norton::HashMap`
* rename `Norton::Objects::HashMap` to `Norton::Objects::Hash`
